patch 7.4.2125 Problem:    Compiler warning for loss of data. Solution:   Add a type cast. (Christian Brabandt)

+++ src/message.c
@@ -304,7 +304,7 @@ trunc_string(
 	    if (len + n > room || half == 0)
 		break;
 	    len += n;
-	    i = half;
+	    i = (int)half;
 	}
     }
     else
